take 1008

+++ Src/Xt.Synth0.DSP/Xts.hpp
@@ -44,11 +44,11 @@ struct XTS_ALIGN PlotState
   int32_t sampleCount, hSplitCount, vSplitCount;
   float* samples;
   int32_t* hSplits;
-  int32_t* vSplits;
+  float* vSplits;
   Xts::SynthModel const* synth;
   std::vector<float>* sampleData;
   std::vector<int32_t>* hSplitData;
-  std::vector<int32_t>* vSplitData;
+  std::vector<float>* vSplitData;
   std::vector<std::complex<float>>* fftData;
   std::vector<std::complex<float>>* fftScratch;
 public:
